hdmitx: update physical size to short int [1/1]
PD#SWPL-121748

Problem:
physical size parsed from edid error

Solution:
Size value is bigger than 8bit, update physical size
to short int.

Verify:
s5

+++ include/linux/amlogic/media/vout/hdmi_tx21/hdmi_tx_module.h
@@ -162,8 +162,8 @@ struct rx_cap {
 	u8 ReceiverProductName[16];
 	u8 manufacture_week;
 	u8 manufacture_year;
-	u8 physical_width;
-	u8 physical_height;
+	u16 physical_width;
+	u16 physical_height;
 	u8 edid_version;
 	u8 edid_revision;
 	u8 ColorDeepSupport;
